feat: move recovery page out of storage wizard
