+++ apps/ubuntu_bootstrap/lib/pages.dart
@@ -6,6 +6,7 @@ export 'pages/autoinstall/autoinstall_page.dart';
 export 'pages/confirm/confirm_page.dart';
 export 'pages/install/install_page.dart';
 export 'pages/loading/loading_page.dart';
+export 'pages/recovery_key/recovery_key_page.dart';
 export 'pages/refresh/refresh_page.dart';
 export 'pages/rst/rst_page.dart';
 export 'pages/secure_boot/secure_boot_page.dart';
