+++ apps/ubuntu_bootstrap/test/recovery_key/recovery_key_page_test.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:ubuntu_bootstrap/pages/storage/recovery_key/recovery_key_model.dart';
-import 'package:ubuntu_bootstrap/pages/storage/recovery_key/recovery_key_page.dart';
+import 'package:ubuntu_bootstrap/pages/recovery_key/recovery_key_model.dart';
+import 'package:ubuntu_bootstrap/pages/recovery_key/recovery_key_page.dart';
 import 'package:ubuntu_provision/ubuntu_provision.dart';
 
+import '../test_utils.dart';
 import 'test_recovery_key.dart';
 
 void main() {
