+++ apps/ubuntu_bootstrap/test/recovery_key/test_recovery_key.dart
@@ -1,9 +1,8 @@
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
-import 'package:ubuntu_bootstrap/pages/storage/recovery_key/recovery_key_model.dart';
+import 'package:ubuntu_bootstrap/pages/recovery_key/recovery_key_model.dart';
 
 import 'test_recovery_key.mocks.dart';
-export '../test_storage.dart';
 export 'test_recovery_key.mocks.dart';
 
 @GenerateMocks([RecoveryKeyModel])
