+++ apps/ubuntu_bootstrap/test/recovery_key/test_recovery_key.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.4 from annotations
-// in ubuntu_bootstrap/test/storage/recovery_key/test_recovery_key.dart.
+// in ubuntu_bootstrap/test/recovery_key/test_recovery_key.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
@@ -7,7 +7,7 @@ import 'dart:async' as _i3;
 import 'dart:ui' as _i4;
 
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:ubuntu_bootstrap/pages/storage/recovery_key/recovery_key_model.dart'
+import 'package:ubuntu_bootstrap/pages/recovery_key/recovery_key_model.dart'
     as _i2;
 
 // ignore_for_file: type=lint
