+++ apps/ubuntu_bootstrap/test/storage/test_storage.dart
@@ -11,7 +11,6 @@ export 'guided_reformat/test_guided_reformat.dart';
 export 'guided_resize/test_guided_resize.dart';
 export 'manual/test_manual_storage.dart';
 export 'passphrase/test_passphrase.dart';
-export 'recovery_key/test_recovery_key.dart';
 export 'test_storage.mocks.dart';
 
 @GenerateMocks([StorageModel])
